requirements.txt
ipapy
markdown-full-yaml-metadata
markdown-checklist
markdown-emdash
mdx-include
pdfkit
pymdown-extensions
